BlockChain refactoring
