+++ test/TestHelper.h
@@ -44,7 +44,6 @@ void mine(Client& c, int numBlocks);
 void connectClients(Client& c1, Client& c2);
 void mine(Block& _s, BlockChain const& _bc);
 void mine(Ethash::BlockHeader& _bi);
-
 }
 
 namespace test
