+++ test/libethereum/blockchain.cpp
@@ -14,20 +14,22 @@
 	You should have received a copy of the GNU General Public License
 	along with cpp-ethereum.  If not, see <http://www.gnu.org/licenses/>.
 */
-/** @file block.cpp
- * @author Christoph Jentzsch <cj@ethdev.com>
+/** @file blockchain.cpp
+ * @author Dimitry Khokhlov <dimitry@ethdev.com>
  * @date 2015
- * block test functions.
+ * BlockChain test functions.
  */
-#include "test/fuzzTesting/fuzzHelper.h"
+#include <boost/filesystem/operations.hpp>
+#include <boost/test/unit_test.hpp>
 #include <boost/filesystem.hpp>
-#include <libdevcore/FileSystem.h>
-#include <libdevcore/TransientDirectory.h>
-#include <libethcore/Params.h>
-#include <libethereum/CanonBlockChain.h>
-#include <libethereum/TransactionQueue.h>
+
 #include <test/TestHelper.h>
-#include <libethereum/Block.h>
+#include <test/BlockChainHelper.h>
+#include <test/JsonSpiritHeaders.h>
+#include "test/fuzzTesting/fuzzHelper.h"
+
+#include <libdevcore/FileSystem.h>
+#include <libethcore/Params.h>
 
 using namespace std;
 using namespace json_spirit;
@@ -36,666 +38,298 @@ using namespace dev::eth;
 
 namespace dev {  namespace test {
 
-typedef std::vector<bytes> uncleList;
-typedef std::pair<bytes, uncleList> blockSet;
-
-bytes createBlockRLPFromFields(mObject& _tObj, h256 const&  _stateRoot = h256{});
-using BlockHeader = Ethash::BlockHeader;
-BlockHeader constructBlock(mObject& _o, h256 const& _stateRoot = h256{});
-RLPStream createFullBlockFromHeader(BlockHeader const& _bi, bytes const& _txs = RLPEmptyList, bytes const& _uncles = RLPEmptyList);
-mArray writeTransactionsToJson(Transactions const& _txs);
-mObject writeBlockHeaderToJson(mObject& _o, BlockHeader const& _bi);
-mArray importUncles(mObject const& _blObj, vector<BlockHeader>& _vBiUncles, vector<BlockHeader> const& _vBiBlocks, std::vector<blockSet> _blockSet);
-void overwriteBlockHeader(BlockHeader& _current_BlockHeader, mObject& _blObj, const BlockHeader& _parent);
-void updatePoW(BlockHeader& _bi);
+//Functions that working with test json
+void compareBlocks(TestBlock const& _a, TestBlock const& _b);
+mArray writeTransactionsToJson(TransactionQueue const& _txsQueue);
+mObject writeBlockHeaderToJson(Ethash::BlockHeader const& _bi);
+void overwriteBlockHeaderForTest(mObject const& _blObj, TestBlock& _block, std::vector<TestBlock> const& importedBlocks);
+void overwriteUncleHeaderForTest(mObject& _uncleHeaderObj, TestBlock& _uncle, std::vector<TestBlock> const& _uncles, std::vector<TestBlock> const& _importedBlocks);
+void checkJsonSectionForInvalidBlock(mObject& _blObj);
+void checkBlocks(TestBlock const& _blockFromFields, TestBlock const& _blockFromRlp, string const& _testname);
 
 void doBlockchainTests(json_spirit::mValue& _v, bool _fillin)
 {
-	string testname;
+	TestOutputHelper::initTest(_v);
 	for (auto& i: _v.get_obj())
 	{
-		mObject& o = i.second.get_obj();
-		if (test::Options::get().singleTest && test::Options::get().singleTestName != i.first)
-		{
-			o.clear();
-			continue;
-		}
+		string testname = i.first;
+		json_spirit::mObject& o = i.second.get_obj();
 
-		cnote << i.first;
-		testname = "(" + i.first + ") ";
+		if (!TestOutputHelper::passTest(o, testname))
+			continue;
+
 		BOOST_REQUIRE(o.count("genesisBlockHeader"));
 		BOOST_REQUIRE(o.count("pre"));
 
-		ImportTest importer(o, _fillin, testType::BlockChainTests);
-		TransientDirectory tdStateDB;
-		BlockHeader biGenesisBlock = constructBlock(o["genesisBlockHeader"].get_obj(), h256{});
-
-		State trueState(OverlayDB(State::openDB(tdStateDB.path(), h256{}, WithExisting::Kill)), BaseState::Empty);
-		ImportTest::importState(o["pre"].get_obj(), trueState);
-		o["pre"] = fillJsonWithState(trueState); //convert all fields to hex
-		trueState.commit();
-
-		//Imported blocks from the start
-		vector<blockSet> blockSets;  //Block(bytes) => UncleList(Blocks(bytes))
-
+		TestBlock genesisBlock(o["genesisBlockHeader"].get_obj(), o["pre"].get_obj());
 		if (_fillin)
-			biGenesisBlock = constructBlock(o["genesisBlockHeader"].get_obj(), trueState.rootHash());
-		else
-			BOOST_CHECK_MESSAGE(biGenesisBlock.stateRoot() == trueState.rootHash(), testname + "root hash does not match");
+			genesisBlock.setBlockHeader(genesisBlock.getBlockHeader(), RecalcBlockHeader::UpdateAndVerify); //update PoW
+		TestBlockChain trueBc(genesisBlock);
 
 		if (_fillin)
 		{
-			// find new valid nonce
-			updatePoW(biGenesisBlock);
-
-			//update genesis block in json file
-			writeBlockHeaderToJson(o["genesisBlockHeader"].get_obj(), biGenesisBlock);
-		}
-
-		// create new "genesis" block
-		RLPStream rlpGenesisBlock = createFullBlockFromHeader(biGenesisBlock);
-		biGenesisBlock.verifyInternals(&rlpGenesisBlock.out());
-		o["genesisRLP"] = toHex(rlpGenesisBlock.out(), 2, HexPrefix::Add);
-
-		// construct true blockchain
-		TransientDirectory td;
-		FullBlockChain<Ethash> trueBc(rlpGenesisBlock.out(), AccountMap(), td.path(), WithExisting::Kill);
-
-		if (_fillin)
-		{
+			o["genesisBlockHeader"] = writeBlockHeaderToJson(genesisBlock.getBlockHeader());
+			o["genesisRLP"] = toHex(genesisBlock.getBytes(), 2, HexPrefix::Add);
 			BOOST_REQUIRE(o.count("blocks"));
+
 			mArray blArray;
-
-			blockSet genesis;
-			genesis.first = rlpGenesisBlock.out();
-			genesis.second = uncleList();
-			blockSets.push_back(genesis);
-			vector<BlockHeader> vBiBlocks;
-			vBiBlocks.push_back(biGenesisBlock);
-
 			size_t importBlockNumber = 0;
+			vector<TestBlock> importedBlocks;
+			importedBlocks.push_back(genesisBlock);
+
 			for (auto const& bl: o["blocks"].get_array())
 			{
+				TestBlockChain blockchain(importedBlocks.at(0));
 				mObject blObj = bl.get_obj();
 				if (blObj.count("blocknumber") > 0)
 					importBlockNumber = max((int)toInt(blObj["blocknumber"]), 1);
 				else
 					importBlockNumber++;
 
-				//each time construct a new blockchain up to importBlockNumber (to generate next block header)
-				vBiBlocks.clear();
-				vBiBlocks.push_back(biGenesisBlock);
+				//Restore blockchain up to block.number to start import from it
+				for (size_t i = 1; i < importedBlocks.size(); i++) //0 block is genesis
+					if (i < importBlockNumber)
+						blockchain.addBlock(importedBlocks.at(i));
+					else
+						break;
 
-				TransientDirectory tdBc;
-				TransientDirectory tdStateDB;
-				FullBlockChain<Ethash> bc(rlpGenesisBlock.out(), AccountMap(), tdBc.path(), WithExisting::Kill);
+				//Drop old blocks
+				size_t originalSize = importedBlocks.size();
+				for (size_t i = importBlockNumber; i < originalSize; i++)
+					importedBlocks.pop_back();
 
-				//OverlayDB database (State::openDB(td_stateDB.path(), h256{}, WithExisting::Kill));
-				State state(OverlayDB(State::openDB(tdStateDB.path(), h256{}, WithExisting::Kill)), BaseState::Empty); //= importer.m_statePre;
-				ImportTest::importState(o["pre"].get_obj(), state);
-				state.commit();
+				TestBlock block;
 
-				//import previous blocks
-				for (size_t i = 1; i < importBlockNumber; i++) //0 block is genesis
-				{
-					BlockQueue uncleQueue;
-					uncleQueue.setChain(bc);
-					uncleList uncles = blockSets.at(i).second;
-					for (size_t j = 0; j < uncles.size(); j++)
-						uncleQueue.import(&uncles.at(j), false);
-
-					const bytes blockFromSet = blockSets.at(i).first;
-					bc.sync(uncleQueue, state.db(), 4);
-					bc.attemptImport(blockFromSet, state.db());
-					vBiBlocks.push_back(BlockHeader(blockFromSet));
-				}
-
-				// get txs
-				TransactionQueue txs;
-				ZeroGasPricer gp;
+				//Import Transactions
 				BOOST_REQUIRE(blObj.count("transactions"));
 				for (auto const& txObj: blObj["transactions"].get_array())
 				{
-					mObject tx = txObj.get_obj();
-					importer.importTransaction(tx);
-					if (txs.import(importer.m_transaction.rlp()) != ImportResult::Success)
-						cnote << "failed importing transaction\n";
+					TestTransaction transaction(txObj.get_obj());
+					block.addTransaction(transaction);
 				}
 
-				//get uncles
-				vector<BlockHeader> vBiUncles;
-				blObj["uncleHeaders"] = importUncles(blObj, vBiUncles, vBiBlocks, blockSets);
-
-				BlockQueue uncleBlockQueue;
-				uncleBlockQueue.setChain(bc);
-				uncleList uncleBlockQueueList;
-				cnote << "import uncle in blockQueue";
-				for (size_t i = 0; i < vBiUncles.size(); i++)
+				//Import Uncles
+				for (auto const& uHObj: blObj.at("uncleHeaders").get_array())
 				{
-					RLPStream uncle = createFullBlockFromHeader(vBiUncles.at(i));
-					try
-					{
-						uncleBlockQueue.import(&uncle.out(), false);
-						uncleBlockQueueList.push_back(uncle.out());
-						// wait until block is verified
-						this_thread::sleep_for(chrono::seconds(1));
-					}
-					catch(...)
-					{
-						cnote << "error in importing uncle! This produces an invalid block (May be by purpose for testing).";
-					}
+					TestBlock uncle;
+					mObject uncleHeaderObj = uHObj.get_obj();
+					overwriteUncleHeaderForTest(uncleHeaderObj, uncle, block.getUncles(), importedBlocks);
+					block.addUncle(uncle);
 				}
 
-				bc.sync(uncleBlockQueue, state.db(), 4);
+				vector<TestBlock> validUncles = blockchain.syncUncles(block.getUncles());
+				block.setUncles(validUncles);
+				block.mine(blockchain);
 
-				Block block = bc.genesisBlock(state.db()); //NOT CLEAR WHAT IT RETURNS IF bc INITIALIZED WITH CUSTOM GENESIS BLOCK
-				block.setBeneficiary(biGenesisBlock.beneficiary());
+				TestBlock alterBlock(block);
+
+				if (blObj.count("blockHeader"))
+					overwriteBlockHeaderForTest(blObj, alterBlock, importedBlocks);
+
+				blObj["rlp"] = toHex(alterBlock.getBytes(), 2, HexPrefix::Add);
+				blObj["blockHeader"] = writeBlockHeaderToJson(alterBlock.getBlockHeader());
+
+				mArray aUncleList;
+				for (size_t i = 0; i < alterBlock.getUncles().size(); i++)
+				{
+					mObject uncleHeaderObj = writeBlockHeaderToJson(alterBlock.getUncles().at(i).getBlockHeader());
+					aUncleList.push_back(uncleHeaderObj);
+				}
+				blObj["uncleHeaders"] = aUncleList;
+				blObj["transactions"] = writeTransactionsToJson(alterBlock.getTransactionQueue());
+
+				bool wasException = false;
+				compareBlocks(block, alterBlock);
 				try
 				{
-					block.sync(bc);
-					block.sync(bc, txs, gp);
-					mine(block, bc);
+					blockchain.addBlock(alterBlock);
+					trueBc.addBlock(alterBlock);
+					importedBlocks.push_back(alterBlock);
 				}
 				catch (Exception const& _e)
 				{
-					cnote << "block sync or mining did throw an exception: " << diagnostic_information(_e);
-					return;
+					cnote << testname + "block import throw an exception: " << diagnostic_information(_e);
+					wasException = true;
 				}
 				catch (std::exception const& _e)
 				{
-					cnote << "block sync or mining did throw an exception: " << _e.what();
-					return;
+					cnote << testname + "block import throw an exception: " << _e.what();
+					wasException = true;
 				}
-
-				blObj["rlp"] = toHex(block.blockData(), 2, HexPrefix::Add);
-
-				//get valid transactions
-				Transactions txList;
-				for (auto const& txi: txs.topTransactions(std::numeric_limits<unsigned>::max()))
-					txList.push_back(txi);
-				blObj["transactions"] = writeTransactionsToJson(txList);
-
-				BlockHeader current_BlockHeader(block.blockData());
-
-				RLPStream uncleStream;
-				uncleStream.appendList(vBiUncles.size());
-				for (unsigned i = 0; i < vBiUncles.size(); ++i)
-				{
-					RLPStream uncleRlp;
-					vBiUncles[i].streamRLP(uncleRlp);
-					uncleStream.appendRaw(uncleRlp.out());
-				}
-
-				if (blObj.count("blockHeader"))
-					overwriteBlockHeader(current_BlockHeader, blObj, bc.info());
-
-				if (blObj.count("blockHeader") && blObj["blockHeader"].get_obj().count("bruncle"))
-					current_BlockHeader.populateFromParent(vBiBlocks[vBiBlocks.size() -1]);
-
-				if (vBiUncles.size())
-				{
-					// update unclehash in case of invalid uncles
-					current_BlockHeader.setSha3Uncles(sha3(uncleStream.out()));
-					updatePoW(current_BlockHeader);
-				}
-
-				// write block header
-				mObject oBlockHeader;
-				writeBlockHeaderToJson(oBlockHeader, current_BlockHeader);
-				blObj["blockHeader"] = oBlockHeader;
-				vBiBlocks.push_back(current_BlockHeader);
-
-				// compare blocks from state and from rlp
-				RLPStream txStream;
-				txStream.appendList(txList.size());
-				for (unsigned i = 0; i < txList.size(); ++i)
-				{
-					RLPStream txrlp;
-					txList[i].streamRLP(txrlp);
-					txStream.appendRaw(txrlp.out());
-				}
-
-				RLPStream blockRLP = createFullBlockFromHeader(current_BlockHeader, txStream.out(), uncleStream.out());
-
-				blObj["rlp"] = toHex(blockRLP.out(), 2, HexPrefix::Add);
-
-				if (sha3(RLP(block.blockData())[0].data()) != sha3(RLP(blockRLP.out())[0].data()))
-				{				
-					cnote << "block header mismatch block.blockData() vs updated block.info()\n";
-					cnote << toHex(RLP(block.blockData())[0].data()) << "vs" << toHex(RLP(blockRLP.out())[0].data());
-				}
-
-				if (sha3(RLP(block.blockData())[1].data()) != sha3(RLP(blockRLP.out())[1].data()))
-					cnote << "txs mismatch\n";
-
-				if (sha3(RLP(block.blockData())[2].data()) != sha3(RLP(blockRLP.out())[2].data()))
-					cnote << "uncle list mismatch\n" << RLP(block.blockData())[2].data() << "\n" << RLP(blockRLP.out())[2].data();
-
-				try
-				{
-					block.sync(bc);
-					bc.import(blockRLP.out(), block.db());
-					block.sync(bc);
-
-					//there we get new blockchain status in state which could have more difficulty than we have in trueState
-					//attempt to import new block to the true blockchain
-					trueBc.sync(uncleBlockQueue, trueState.db(), 4);
-					trueBc.attemptImport(blockRLP.out(), trueState.db());
-
-					if (block.blockData() == trueBc.block())
-						trueState = block.state();
-
-					blockSet newBlock;
-					newBlock.first = blockRLP.out();
-					newBlock.second = uncleBlockQueueList;
-					if (importBlockNumber < blockSets.size())
-					{
-						//make new correct history of imported blocks
-						blockSets[importBlockNumber] = newBlock;
-						for (size_t i = importBlockNumber + 1; i < blockSets.size(); i++)
-							blockSets.pop_back();
-					}
-					else
-						blockSets.push_back(newBlock);
-				}
-				// if exception is thrown, RLP is invalid and no blockHeader, Transaction list, or Uncle list should be given
 				catch (...)
 				{
+					wasException = true;
+				}
+
+				if (wasException)
+				{
+					// if exception is thrown, RLP is invalid and no blockHeader, Transaction list, or Uncle list should be given
 					cnote << "block is invalid!\n";
 					blObj.erase(blObj.find("blockHeader"));
 					blObj.erase(blObj.find("uncleHeaders"));
 					blObj.erase(blObj.find("transactions"));
 				}
-				blArray.push_back(blObj);
+
+				blArray.push_back(blObj);  //json data
 				this_thread::sleep_for(chrono::seconds(1));
-			} //for blocks
+			}//each blocks
 
 			if (o.count("expect") > 0)
 			{
 				AccountMaskMap expectStateMap;
 				State stateExpect(OverlayDB(), BaseState::Empty);
 				ImportTest::importState(o["expect"].get_obj(), stateExpect, expectStateMap);
-				if (ImportTest::compareStates(stateExpect, trueState, expectStateMap, Options::get().checkState ? WhenError::Throw : WhenError::DontThrow))
+				if (ImportTest::compareStates(stateExpect, trueBc.getTopBlock().getState(), expectStateMap, Options::get().checkState ? WhenError::Throw : WhenError::DontThrow))
 					cerr << testname << endl;
 				o.erase(o.find("expect"));
 			}
 
 			o["blocks"] = blArray;
-			o["postState"] = fillJsonWithState(trueState);
-			o["lastblockhash"] = toString(trueBc.info().hash());
+			o["postState"] = fillJsonWithState(trueBc.getTopBlock().getState());
+			o["lastblockhash"] = toString(trueBc.getTopBlock().getBlockHeader().hash());
 
 			//make all values hex in pre section
 			State prestate(OverlayDB(), BaseState::Empty);
 			ImportTest::importState(o["pre"].get_obj(), prestate);
 			o["pre"] = fillJsonWithState(prestate);
-		}//_fillin
+		}//fillin
 		else
 		{
+			TestBlockChain blockchain(genesisBlock);
 			for (auto const& bl: o["blocks"].get_array())
 			{
-				bool importedAndBest = true;
 				mObject blObj = bl.get_obj();
-				bytes blockRLP;
+				TestBlock blockFromRlp;
 				try
 				{
-					blockRLP = importByteArray(blObj["rlp"].get_str());
-					trueBc.import(blockRLP, trueState.db());
-					if (trueBc.info() != BlockHeader(blockRLP))
-						importedAndBest  = false;
+					TestBlock blRlp(blObj["rlp"].get_str());
+					blockFromRlp = blRlp;
+					trueBc.addBlock(blRlp);
 				}
 				// if exception is thrown, RLP is invalid and no blockHeader, Transaction list, or Uncle list should be given
 				catch (Exception const& _e)
 				{
 					cnote << "state sync or block import did throw an exception: " << diagnostic_information(_e);
-					BOOST_CHECK(blObj.count("blockHeader") == 0);
-					BOOST_CHECK(blObj.count("transactions") == 0);
-					BOOST_CHECK(blObj.count("uncleHeaders") == 0);
+					checkJsonSectionForInvalidBlock(blObj);
 					continue;
 				}
 				catch (std::exception const& _e)
 				{
 					cnote << "state sync or block import did throw an exception: " << _e.what();
-					BOOST_CHECK(blObj.count("blockHeader") == 0);
-					BOOST_CHECK(blObj.count("transactions") == 0);
-					BOOST_CHECK(blObj.count("uncleHeaders") == 0);
+					checkJsonSectionForInvalidBlock(blObj);
 					continue;
 				}
 				catch (...)
 				{
 					cnote << "state sync or block import did throw an exception\n";
-					BOOST_CHECK(blObj.count("blockHeader") == 0);
-					BOOST_CHECK(blObj.count("transactions") == 0);
-					BOOST_CHECK(blObj.count("uncleHeaders") == 0);
+					checkJsonSectionForInvalidBlock(blObj);
 					continue;
 				}
 
+				//block from RLP successfully imported. now compare this rlp to test sections
 				BOOST_REQUIRE(blObj.count("blockHeader"));
 
-				mObject tObj = blObj["blockHeader"].get_obj();
-				BlockHeader blockHeaderFromFields;
-				const bytes c_rlpBytesBlockHeader = createBlockRLPFromFields(tObj);
-				const RLP c_blockHeaderRLP(c_rlpBytesBlockHeader);
-				blockHeaderFromFields.populateFromHeader(c_blockHeaderRLP, IgnoreSeal);
+				//Check Provided Header against block in RLP
+				mObject emptyState;
+				TestBlock blockFromFields(blObj["blockHeader"].get_obj(), emptyState, RecalcBlockHeader::SkipVerify); //not verify as we havent imported transactions and uncles yet
 
-				BlockHeader blockFromRlp(trueBc.header());
-
-				if (importedAndBest)
+				//ImportTransactions
+				BOOST_REQUIRE(blObj.count("transactions"));
+				for (auto const& txObj: blObj["transactions"].get_array())
 				{
-					//Check the fields restored from RLP to original fields
-					BOOST_CHECK_MESSAGE(blockHeaderFromFields.headerHash(WithProof) == blockFromRlp.headerHash(WithProof), testname + "hash in given RLP not matching the block hash!");
-					BOOST_CHECK_MESSAGE(blockHeaderFromFields.parentHash() == blockFromRlp.parentHash(), testname + "parentHash in given RLP not matching the block parentHash!");
-					BOOST_CHECK_MESSAGE(blockHeaderFromFields.sha3Uncles() == blockFromRlp.sha3Uncles(), testname + "sha3Uncles in given RLP not matching the block sha3Uncles!");
-					BOOST_CHECK_MESSAGE(blockHeaderFromFields.beneficiary() == blockFromRlp.beneficiary(), testname + "beneficiary in given RLP not matching the block beneficiary!");
-					BOOST_CHECK_MESSAGE(blockHeaderFromFields.stateRoot() == blockFromRlp.stateRoot(), testname + "stateRoot in given RLP not matching the block stateRoot!");
-					BOOST_CHECK_MESSAGE(blockHeaderFromFields.transactionsRoot() == blockFromRlp.transactionsRoot(), testname + "transactionsRoot in given RLP not matching the block transactionsRoot!");
-					BOOST_CHECK_MESSAGE(blockHeaderFromFields.receiptsRoot() == blockFromRlp.receiptsRoot(), testname + "receiptsRoot in given RLP not matching the block receiptsRoot!");
-					BOOST_CHECK_MESSAGE(blockHeaderFromFields.logBloom() == blockFromRlp.logBloom(), testname + "logBloom in given RLP not matching the block logBloom!");
-					BOOST_CHECK_MESSAGE(blockHeaderFromFields.difficulty() == blockFromRlp.difficulty(), testname + "difficulty in given RLP not matching the block difficulty!");
-					BOOST_CHECK_MESSAGE(blockHeaderFromFields.number() == blockFromRlp.number(), testname + "number in given RLP not matching the block number!");
-					BOOST_CHECK_MESSAGE(blockHeaderFromFields.gasLimit() == blockFromRlp.gasLimit(),"testname + gasLimit in given RLP not matching the block gasLimit!");
-					BOOST_CHECK_MESSAGE(blockHeaderFromFields.gasUsed() == blockFromRlp.gasUsed(), testname + "gasUsed in given RLP not matching the block gasUsed!");
-					BOOST_CHECK_MESSAGE(blockHeaderFromFields.timestamp() == blockFromRlp.timestamp(), testname + "timestamp in given RLP not matching the block timestamp!");
-					BOOST_CHECK_MESSAGE(blockHeaderFromFields.extraData() == blockFromRlp.extraData(), testname + "extraData in given RLP not matching the block extraData!");
-					BOOST_CHECK_MESSAGE(blockHeaderFromFields.mixHash() == blockFromRlp.mixHash(), testname + "mixHash in given RLP not matching the block mixHash!");
-					BOOST_CHECK_MESSAGE(blockHeaderFromFields.nonce() == blockFromRlp.nonce(), testname + "nonce in given RLP not matching the block nonce!");
+					TestTransaction transaction(txObj.get_obj());
+					blockFromFields.addTransaction(transaction);
+				}
 
-					BOOST_CHECK_MESSAGE(blockHeaderFromFields == blockFromRlp, testname + "However, blockHeaderFromFields != blockFromRlp!");
-
-					//Check transaction list
-					Transactions txsFromField;
-
-					for (auto const& txObj: blObj["transactions"].get_array())
+				// ImportUncles
+				if (blObj["uncleHeaders"].type() != json_spirit::null_type)
+					for (auto const& uBlHeaderObj: blObj["uncleHeaders"].get_array())
 					{
-						mObject tx = txObj.get_obj();
+						mObject uBlH = uBlHeaderObj.get_obj();
+						BOOST_REQUIRE((uBlH.size() == 16));
 
-						BOOST_REQUIRE(tx.count("nonce"));
-						BOOST_REQUIRE(tx.count("gasPrice"));
-						BOOST_REQUIRE(tx.count("gasLimit"));
-						BOOST_REQUIRE(tx.count("to"));
-						BOOST_REQUIRE(tx.count("value"));
-						BOOST_REQUIRE(tx.count("v"));
-						BOOST_REQUIRE(tx.count("r"));
-						BOOST_REQUIRE(tx.count("s"));
-						BOOST_REQUIRE(tx.count("data"));
-
-						try
-						{
-							Transaction t(createRLPStreamFromTransactionFields(tx).out(), CheckTransaction::Everything);
-							txsFromField.push_back(t);
-						}
-						catch (Exception const& _e)
-						{
-							BOOST_ERROR("Failed transaction constructor with Exception: " << diagnostic_information(_e));
-						}
-						catch (exception const& _e)
-						{
-							cnote << _e.what();
-						}
+						mObject emptyState;
+						TestBlock uncle(uBlH, emptyState);
+						blockFromFields.addUncle(uncle);
 					}
 
-					Transactions txsFromRlp;
-					RLP root(blockRLP);
-					for (auto const& tr: root[1])
-					{
-						Transaction tx(tr.data(), CheckTransaction::Everything);
-						txsFromRlp.push_back(tx);
-					}
+				checkBlocks(blockFromFields, blockFromRlp, testname);
 
-					BOOST_CHECK_MESSAGE(txsFromRlp.size() == txsFromField.size(), "transaction list size does not match");
+				try
+				{
+					blockFromFields.setBlockHeader(blockFromFields.getBlockHeader(), RecalcBlockHeader::Verify); //recalculateBytes
+					blockchain.addBlock(blockFromFields);
+				}
+				catch (Exception const& _e)
+				{
+					cerr << "Error importing block from filds to blockchain: " << diagnostic_information(_e);
+					break;
+				}
+			}//allBlocks
 
-					for (size_t i = 0; i < txsFromField.size(); ++i)
-					{
-						BOOST_CHECK_MESSAGE(txsFromField[i].data() == txsFromRlp[i].data(), testname + "transaction data in rlp and in field do not match");
-						BOOST_CHECK_MESSAGE(txsFromField[i].gas() == txsFromRlp[i].gas(), testname + "transaction gasLimit in rlp and in field do not match");
-						BOOST_CHECK_MESSAGE(txsFromField[i].gasPrice() == txsFromRlp[i].gasPrice(), testname + "transaction gasPrice in rlp and in field do not match");
-						BOOST_CHECK_MESSAGE(txsFromField[i].nonce() == txsFromRlp[i].nonce(), testname + "transaction nonce in rlp and in field do not match");
-						BOOST_CHECK_MESSAGE(txsFromField[i].signature().r == txsFromRlp[i].signature().r, testname + "transaction r in rlp and in field do not match");
-						BOOST_CHECK_MESSAGE(txsFromField[i].signature().s == txsFromRlp[i].signature().s, testname + "transaction s in rlp and in field do not match");
-						BOOST_CHECK_MESSAGE(txsFromField[i].signature().v == txsFromRlp[i].signature().v, testname + "transaction v in rlp and in field do not match");
-						BOOST_CHECK_MESSAGE(txsFromField[i].receiveAddress() == txsFromRlp[i].receiveAddress(), testname + "transaction receiveAddress in rlp and in field do not match");
-						BOOST_CHECK_MESSAGE(txsFromField[i].value() == txsFromRlp[i].value(), testname + "transaction receiveAddress in rlp and in field do not match");
+			BOOST_REQUIRE((o.count("lastblockhash") > 0));
+			string lastTrueBlockHash = toString(trueBc.getTopBlock().getBlockHeader().hash());
+			BOOST_CHECK_MESSAGE(lastTrueBlockHash == o["lastblockhash"].get_str(),
+					testname + "Boost check: lastblockhash does not match " + lastTrueBlockHash + " expected: " + o["lastblockhash"].get_str());
 
-						BOOST_CHECK_MESSAGE(txsFromField[i] == txsFromRlp[i], testname + "transactions from  rlp and transaction from field do not match");
-						BOOST_CHECK_MESSAGE(txsFromField[i].rlp() == txsFromRlp[i].rlp(), testname + "transactions rlp do not match");
-					}
+			BOOST_CHECK_MESSAGE(toString(trueBc.getTopBlock().getState().rootHash()) ==
+								toString(blockchain.getTopBlock().getState().rootHash()),
+								testname + "State root in chain from RLP blocks != State root in chain from Field blocks!");
 
-					// check uncle list
-
-					// uncles from uncle list field
-					vector<BlockHeader> uBlHsFromField;
-					if (blObj["uncleHeaders"].type() != json_spirit::null_type)
-						for (auto const& uBlHeaderObj: blObj["uncleHeaders"].get_array())
-						{
-							mObject uBlH = uBlHeaderObj.get_obj();
-							BOOST_REQUIRE(uBlH.size() == 16);
-							bytes uncleRLP = createBlockRLPFromFields(uBlH);
-							const RLP c_uRLP(uncleRLP);
-							BlockHeader uncleBlockHeader;
-							try
-							{
-								uncleBlockHeader.populateFromHeader(c_uRLP);
-							}
-							catch(...)
-							{
-								BOOST_ERROR(testname + "invalid uncle header");
-							}
-							uBlHsFromField.push_back(uncleBlockHeader);
-						}
-
-					// uncles from block RLP
-					vector<BlockHeader> uBlHsFromRlp;
-					for	(auto const& uRLP: root[2])
-					{
-						BlockHeader uBl;
-						uBl.populateFromHeader(uRLP);
-						uBlHsFromRlp.push_back(uBl);
-					}
-
-					BOOST_REQUIRE_EQUAL(uBlHsFromField.size(), uBlHsFromRlp.size());
-
-					for (size_t i = 0; i < uBlHsFromField.size(); ++i)
-						BOOST_CHECK_MESSAGE(uBlHsFromField[i] == uBlHsFromRlp[i], testname + "block header in rlp and in field do not match");
-				}//importedAndBest
-			}//all blocks
-
-			BOOST_REQUIRE(o.count("lastblockhash") > 0);
-			BOOST_CHECK_MESSAGE(toString(trueBc.info().hash()) == o["lastblockhash"].get_str(),
-					testname + "Boost check: " + i.first + " lastblockhash does not match " + toString(trueBc.info().hash()) + " expected: " + o["lastblockhash"].get_str());
+			State postState; //Compare post states
+			BOOST_REQUIRE((o.count("postState") > 0));
+			ImportTest::importState(o["postState"].get_obj(), postState);
+			ImportTest::compareStates(postState, trueBc.getTopBlock().getState());
+			ImportTest::compareStates(postState, blockchain.getTopBlock().getState());
 		}
-	}
-}
-// helping functions
-
-mArray importUncles(mObject const& _blObj, vector<BlockHeader>& _vBiUncles, vector<BlockHeader> const& _vBiBlocks, std::vector<blockSet> _blockSet)
-{
-	// write uncle list
-	mArray aUncleList;
-	mObject uncleHeaderObj_pre;
-
-	for (auto const& uHObj: _blObj.at("uncleHeaders").get_array())
-	{
-		mObject uncleHeaderObj = uHObj.get_obj();
-		if (uncleHeaderObj.count("sameAsPreviousSibling"))
-		{
-			writeBlockHeaderToJson(uncleHeaderObj_pre, _vBiUncles[_vBiUncles.size()-1]);
-			aUncleList.push_back(uncleHeaderObj_pre);
-			_vBiUncles.push_back(_vBiUncles[_vBiUncles.size()-1]);
-			uncleHeaderObj_pre = uncleHeaderObj;
-			continue;
-		}
-
-		if (uncleHeaderObj.count("sameAsBlock"))
-		{
-			size_t number = (size_t)toInt(uncleHeaderObj["sameAsBlock"]);
-			uncleHeaderObj.erase("sameAsBlock");
-			BlockHeader currentUncle = _vBiBlocks[number];
-			writeBlockHeaderToJson(uncleHeaderObj, currentUncle);
-			aUncleList.push_back(uncleHeaderObj);
-			_vBiUncles.push_back(currentUncle);
-			uncleHeaderObj_pre = uncleHeaderObj;
-			continue;
-		}
-
-		if (uncleHeaderObj.count("sameAsPreviousBlockUncle"))
-		{
-			bytes uncleRLP = _blockSet[(size_t)toInt(uncleHeaderObj["sameAsPreviousBlockUncle"])].second[0];
-			BlockHeader uncleHeader(uncleRLP);
-			writeBlockHeaderToJson(uncleHeaderObj, uncleHeader);
-			aUncleList.push_back(uncleHeaderObj);
-
-			_vBiUncles.push_back(uncleHeader);
-			uncleHeaderObj_pre = uncleHeaderObj;
-			continue;
-		}
-
-		string overwrite = "false";
-		if (uncleHeaderObj.count("overwriteAndRedoPoW"))
-		{
-			overwrite = uncleHeaderObj["overwriteAndRedoPoW"].get_str();
-			uncleHeaderObj.erase("overwriteAndRedoPoW");
-		}
-
-		BlockHeader uncleBlockFromFields = constructBlock(uncleHeaderObj);
-
-		// make uncle header valid
-		uncleBlockFromFields.setTimestamp(utcTime());
-		if (_vBiBlocks.size() > 2)
-		{
-			if (uncleBlockFromFields.number() - 1 < _vBiBlocks.size())
-				uncleBlockFromFields.populateFromParent(_vBiBlocks[(size_t)uncleBlockFromFields.number() - 1]);
-			else
-				uncleBlockFromFields.populateFromParent(_vBiBlocks[_vBiBlocks.size() - 2]);
-		}
-		else
-			continue;
-
-		if (overwrite != "false")
-		{
-			uncleBlockFromFields = constructHeader(
-				overwrite == "parentHash" ? h256(uncleHeaderObj["parentHash"].get_str()) : uncleBlockFromFields.parentHash(),
-				uncleBlockFromFields.sha3Uncles(),
-				uncleBlockFromFields.beneficiary(),
-				overwrite == "stateRoot" ? h256(uncleHeaderObj["stateRoot"].get_str()) : uncleBlockFromFields.stateRoot(),
-				uncleBlockFromFields.transactionsRoot(),
-				uncleBlockFromFields.receiptsRoot(),
-				uncleBlockFromFields.logBloom(),
-				overwrite == "difficulty" ? toInt(uncleHeaderObj["difficulty"]) : overwrite == "timestamp" ? uncleBlockFromFields.calculateDifficulty(_vBiBlocks[(size_t)uncleBlockFromFields.number() - 1]) : uncleBlockFromFields.difficulty(),
-				uncleBlockFromFields.number(),
-				overwrite == "gasLimit" ? toInt(uncleHeaderObj["gasLimit"]) : uncleBlockFromFields.gasLimit(),
-				overwrite == "gasUsed" ? toInt(uncleHeaderObj["gasUsed"]) : uncleBlockFromFields.gasUsed(),
-				overwrite == "timestamp" ? toInt(uncleHeaderObj["timestamp"]) : uncleBlockFromFields.timestamp(),
-				uncleBlockFromFields.extraData());
-
-			if (overwrite == "parentHashIsBlocksParent")
-				uncleBlockFromFields.populateFromParent(_vBiBlocks[_vBiBlocks.size() - 1]);
-		}
-
-
-		updatePoW(uncleBlockFromFields);
-
-		if (overwrite == "nonce")
-			updateEthashSeal(uncleBlockFromFields, uncleBlockFromFields.mixHash(), Nonce(uncleHeaderObj["nonce"].get_str()));
-
-		if (overwrite == "mixHash")
-			updateEthashSeal(uncleBlockFromFields, h256(uncleHeaderObj["mixHash"].get_str()), uncleBlockFromFields.nonce());
-
-		writeBlockHeaderToJson(uncleHeaderObj, uncleBlockFromFields);
-
-		aUncleList.push_back(uncleHeaderObj);
-		_vBiUncles.push_back(uncleBlockFromFields);
-
-		uncleHeaderObj_pre = uncleHeaderObj;
-	} //for _blObj["uncleHeaders"].get_array()
-
-	return aUncleList;
+	}//for tests
 }
 
-bytes createBlockRLPFromFields(mObject& _tObj, h256 const& _stateRoot)
+//TestFunction
+void overwriteBlockHeaderForTest(mObject const& _blObj, TestBlock& _block, std::vector<TestBlock> const& _importedBlocks)
 {
-	RLPStream rlpStream;
-	rlpStream.appendList(_tObj.count("hash") > 0 ? (_tObj.size() - 1) : _tObj.size());
+	//_blObj  - json object with header data
+	//_block  - which header would be overwritten
+	//_parentHeader - parent blockheader
 
-	if (_tObj.count("parentHash"))
-		rlpStream << importByteArray(_tObj["parentHash"].get_str());
-
-	if (_tObj.count("uncleHash"))
-		rlpStream << importByteArray(_tObj["uncleHash"].get_str());
-
-	if (_tObj.count("coinbase"))
-		rlpStream << importByteArray(_tObj["coinbase"].get_str());
-
-	if (_stateRoot)
-		rlpStream << _stateRoot;
-	else if (_tObj.count("stateRoot"))
-		rlpStream << importByteArray(_tObj["stateRoot"].get_str());
-
-	if (_tObj.count("transactionsTrie"))
-		rlpStream << importByteArray(_tObj["transactionsTrie"].get_str());
-
-	if (_tObj.count("receiptTrie"))
-		rlpStream << importByteArray(_tObj["receiptTrie"].get_str());
-
-	if (_tObj.count("bloom"))
-		rlpStream << importByteArray(_tObj["bloom"].get_str());
-
-	if (_tObj.count("difficulty"))
-		rlpStream << bigint(_tObj["difficulty"].get_str());
-
-	if (_tObj.count("number"))
-		rlpStream << bigint(_tObj["number"].get_str());
-
-	if (_tObj.count("gasLimit"))
-		rlpStream << bigint(_tObj["gasLimit"].get_str());
-
-	if (_tObj.count("gasUsed"))
-		rlpStream << bigint(_tObj["gasUsed"].get_str());
-
-	if (_tObj.count("timestamp"))
-		rlpStream << bigint(_tObj["timestamp"].get_str());
-
-	if (_tObj.count("extraData"))
-		rlpStream << fromHex(_tObj["extraData"].get_str());
-
-	if (_tObj.count("mixHash"))
-		rlpStream << importByteArray(_tObj["mixHash"].get_str());
-
-	if (_tObj.count("nonce"))
-		rlpStream << importByteArray(_tObj["nonce"].get_str());
-
-	return rlpStream.out();
-}
-
-void overwriteBlockHeader(BlockHeader& _header, mObject& _blObj, BlockHeader const& _parent)
-{
-	auto ho = _blObj["blockHeader"].get_obj();
+	RecalcBlockHeader findNewValidNonce = RecalcBlockHeader::Verify;
+	Ethash::BlockHeader tmp;
+	Ethash::BlockHeader const& header = _block.getBlockHeader();
+	auto ho = _blObj.at("blockHeader").get_obj();
 	if (ho.size() != 14)
 	{
-		BlockHeader tmp = constructHeader(
-			ho.count("parentHash") ? h256(ho["parentHash"].get_str()) : _header.parentHash(),
-			ho.count("uncleHash") ? h256(ho["uncleHash"].get_str()) : _header.sha3Uncles(),
-			ho.count("coinbase") ? Address(ho["coinbase"].get_str()) : _header.beneficiary(),
-			ho.count("stateRoot") ? h256(ho["stateRoot"].get_str()): _header.stateRoot(),
-			ho.count("transactionsTrie") ? h256(ho["transactionsTrie"].get_str()) : _header.transactionsRoot(),
-			ho.count("receiptTrie") ? h256(ho["receiptTrie"].get_str()) : _header.receiptsRoot(),
-			ho.count("bloom") ? LogBloom(ho["bloom"].get_str()) : _header.logBloom(),
-			ho.count("difficulty") ? toInt(ho["difficulty"]) : _header.difficulty(),
-			ho.count("number") ? toInt(ho["number"]) : _header.number(),
-			ho.count("gasLimit") ? toInt(ho["gasLimit"]) : _header.gasLimit(),
-			ho.count("gasUsed") ? toInt(ho["gasUsed"]) : _header.gasUsed(),
-			ho.count("timestamp") ? toInt(ho["timestamp"]) : _header.timestamp(),
-			ho.count("extraData") ? importByteArray(ho["extraData"].get_str()) : _header.extraData());
+		tmp = constructHeader(
+			ho.count("parentHash") ? h256(ho["parentHash"].get_str()) : header.parentHash(),
+			ho.count("uncleHash") ? h256(ho["uncleHash"].get_str()) : header.sha3Uncles(),
+			ho.count("coinbase") ? Address(ho["coinbase"].get_str()) : header.beneficiary(),
+			ho.count("stateRoot") ? h256(ho["stateRoot"].get_str()): header.stateRoot(),
+			ho.count("transactionsTrie") ? h256(ho["transactionsTrie"].get_str()) : header.transactionsRoot(),
+			ho.count("receiptTrie") ? h256(ho["receiptTrie"].get_str()) : header.receiptsRoot(),
+			ho.count("bloom") ? LogBloom(ho["bloom"].get_str()) : header.logBloom(),
+			ho.count("difficulty") ? toInt(ho["difficulty"]) : header.difficulty(),
+			ho.count("number") ? toInt(ho["number"]) : header.number(),
+			ho.count("gasLimit") ? toInt(ho["gasLimit"]) : header.gasLimit(),
+			ho.count("gasUsed") ? toInt(ho["gasUsed"]) : header.gasUsed(),
+			ho.count("timestamp") ? toInt(ho["timestamp"]) : header.timestamp(),
+			ho.count("extraData") ? importByteArray(ho["extraData"].get_str()) : header.extraData());
 
 		if (ho.count("RelTimestamp"))
 		{
-			tmp.setTimestamp(toInt(ho["RelTimestamp"]) +_parent.timestamp());
-			tmp.setDifficulty(tmp.calculateDifficulty(_parent));
+			Ethash::BlockHeader parentHeader = _importedBlocks.at(_importedBlocks.size() - 1).getBlockHeader();
+			tmp.setTimestamp(toInt(ho["RelTimestamp"]) + parentHeader.timestamp());
+			tmp.setDifficulty(tmp.calculateDifficulty(parentHeader));
 			this_thread::sleep_for(chrono::seconds((int)toInt(ho["RelTimestamp"])));
 		}
 
 		// find new valid nonce
-		if (static_cast<BlockInfo>(tmp) != static_cast<BlockInfo>(_header) && tmp.difficulty())
-			mine(tmp);
+		if (static_cast<BlockInfo>(tmp) != static_cast<BlockInfo>(header) && tmp.difficulty())
+			findNewValidNonce = RecalcBlockHeader::Update;
 
+		if (ho.count("updatePoW"))
+			findNewValidNonce = RecalcBlockHeader::UpdateAndVerify;
+
+		tmp.setNonce(header.nonce());
+		tmp.setMixHash(header.mixHash());
 
 		if (ho.count("mixHash"))
 			updateEthashSeal(tmp, h256(ho["mixHash"].get_str()), tmp.nonce());
@@ -703,52 +337,169 @@ void overwriteBlockHeader(BlockHeader& _header, mObject& _blObj, BlockHeader con
 			updateEthashSeal(tmp, tmp.mixHash(), Nonce(ho["nonce"].get_str()));
 
 		tmp.noteDirty();
-		_header = tmp;
 	}
 	else
 	{
 		// take the blockheader as is
-		const bytes c_blockRLP = createBlockRLPFromFields(ho);
-		const RLP c_bRLP(c_blockRLP);
-		_header.populateFromHeader(c_bRLP, IgnoreSeal);
+		mObject emptyState;
+		tmp = TestBlock(ho, emptyState, RecalcBlockHeader::SkipVerify).getBlockHeader();
 	}
+
+
+	if (ho.count("populateFromBlock"))
+	{
+		size_t number = (size_t)toInt(ho.at("populateFromBlock"));
+		ho.erase("populateFromBlock");
+		if (number < _importedBlocks.size())
+		{
+			Ethash::BlockHeader parentHeader = _importedBlocks.at(number).getBlockHeader();
+			tmp.populateFromParent(parentHeader);
+			findNewValidNonce = RecalcBlockHeader::UpdateAndVerify;
+		}
+		else
+		{
+			cerr << "Could not populate blockHeader from block: there are no block with number!" << TestOutputHelper::testName() << endl;
+		}
+	}
+
+	_block.setBlockHeader(tmp, findNewValidNonce);
 }
 
-BlockHeader constructBlock(mObject& _o, h256 const& _stateRoot)
+void overwriteUncleHeaderForTest(mObject& uncleHeaderObj, TestBlock& uncle, std::vector<TestBlock> const& uncles, std::vector<TestBlock> const& importedBlocks)
 {
-	BlockHeader ret;
-	try
+	//uncleHeaderObj - json Uncle header with additional option fields
+	//uncle			 - uncle Block to overwrite
+	//uncles		 - previously imported uncles
+	//importedBlocks - blocks already included in BlockChain
+
+	if (uncleHeaderObj.count("sameAsPreviousSibling"))
 	{
-		// construct genesis block
-		const bytes c_blockRLP = createBlockRLPFromFields(_o, _stateRoot);
-		const RLP c_bRLP(c_blockRLP);
-		ret.populateFromHeader(c_bRLP, IgnoreSeal);
+		uncleHeaderObj.erase("sameAsPreviousSibling");
+		if (uncles.size())
+			uncle = uncles.at(uncles.size() - 1);
+		else
+			cerr << "Could not create uncle sameAsPreviousSibling: there are no siblings!";
+		return;
 	}
-	catch (Exception const& _e)
+
+	if (uncleHeaderObj.count("sameAsBlock"))
 	{
-		cnote << "block population did throw an exception: " << diagnostic_information(_e);
+		size_t number = (size_t)toInt(uncleHeaderObj.at("sameAsBlock"));
+		uncleHeaderObj.erase("sameAsBlock");
+
+		if (number < importedBlocks.size())
+			uncle = importedBlocks.at(number);
+		else
+			cerr << "Could not create uncle sameAsBlock: there are no block with number " << number;
+		return;
 	}
-	catch (std::exception const& _e)
+
+	if (uncleHeaderObj.count("sameAsPreviousBlockUncle"))
 	{
-		BOOST_ERROR("Failed block population with Exception: " << _e.what());
+		size_t number = (size_t)toInt(uncleHeaderObj.at("sameAsPreviousBlockUncle"));
+		uncleHeaderObj.erase("sameAsPreviousBlockUncle");
+		if (number < importedBlocks.size())
+		{
+			vector<TestBlock> prevBlockUncles = importedBlocks.at(number).getUncles();
+			if (prevBlockUncles.size())
+				uncle = prevBlockUncles[0];  //exact uncle??
+			else
+				cerr << "Could not create uncle sameAsPreviousBlockUncle: previous block has no uncles!" << TestOutputHelper::testName() << endl;
+		}
+		else
+			cerr << "Could not create uncle sameAsPreviousBlockUncle: there are no block imported!" << TestOutputHelper::testName() << endl;
+		return;
 	}
-	catch(...)
+
+	string overwrite = "false";
+	if (uncleHeaderObj.count("overwriteAndRedoPoW"))
 	{
-		BOOST_ERROR("block population did throw an unknown exception\n");
+		overwrite = uncleHeaderObj.at("overwriteAndRedoPoW").get_str();
+		uncleHeaderObj.erase("overwriteAndRedoPoW");
 	}
-	return ret;
+
+	//construct actual block
+	Ethash::BlockHeader uncleHeader;
+	if (uncleHeaderObj.count("populateFromBlock"))
+	{
+		uncleHeader.setTimestamp((u256)time(0));
+		size_t number = (size_t)toInt(uncleHeaderObj.at("populateFromBlock"));
+		uncleHeaderObj.erase("populateFromBlock");
+		if (number < importedBlocks.size())
+		{
+			uncleHeader.populateFromParent(importedBlocks.at(number).getBlockHeader());
+			//Set Default roots for empty block
+			//m_transactionsRoot = _t; m_receiptsRoot = _r; m_sha3Uncles = _u; m_stateRoot = _s;
+			uncleHeader.setRoots((h256)fromHex("56e81f171bcc55a6ff8345e692c0f86e5b48e01b996cadc001622fb5e363b421"),
+								 (h256)fromHex("56e81f171bcc55a6ff8345e692c0f86e5b48e01b996cadc001622fb5e363b421"),
+								 (h256)fromHex("1dcc4de8dec75d7aab85b567b6ccd41ad312451b948a7413f0a142fd40d49347"),
+								 uncleHeader.stateRoot()
+								 );
+		}
+		else
+			cerr << "Could not create uncle populateFromBlock: there are no block with number " << number << TestOutputHelper::testName() << endl;
+	}
+	else
+	{
+		mObject emptyState;
+		uncle = TestBlock(uncleHeaderObj, emptyState, RecalcBlockHeader::SkipVerify);
+		uncleHeader = uncle.getBlockHeader();
+	}
+
+	if (overwrite != "false")
+	{
+		uncleHeader = constructHeader(
+			overwrite == "parentHash" ? h256(uncleHeaderObj.at("parentHash").get_str()) : uncleHeader.parentHash(),
+			uncleHeader.sha3Uncles(),
+			overwrite == "coinbase" ? Address(uncleHeaderObj.at("coinbase").get_str()) : uncleHeader.beneficiary(),
+			overwrite == "stateRoot" ? h256(uncleHeaderObj.at("stateRoot").get_str()) : uncleHeader.stateRoot(),
+			uncleHeader.transactionsRoot(),
+			uncleHeader.receiptsRoot(),
+			uncleHeader.logBloom(),
+			overwrite == "difficulty" ? toInt(uncleHeaderObj.at("difficulty"))
+									  :	overwrite == "timestamp" ? uncleHeader.calculateDifficulty(importedBlocks.at((size_t)uncleHeader.number() - 1).getBlockHeader())
+																 : uncleHeader.difficulty(),
+			overwrite == "number" ? toInt(uncleHeaderObj.at("number")) : uncleHeader.number(),
+			overwrite == "gasLimit" ? toInt(uncleHeaderObj.at("gasLimit")) : uncleHeader.gasLimit(),
+			overwrite == "gasUsed" ? toInt(uncleHeaderObj.at("gasUsed")) : uncleHeader.gasUsed(),
+			overwrite == "timestamp" ? toInt(uncleHeaderObj.at("timestamp")) : uncleHeader.timestamp(),
+			uncleHeader.extraData());
+	}
+
+	if (overwrite == "nonce" || overwrite == "mixHash")
+	{
+		if (overwrite == "nonce")
+			updateEthashSeal(uncleHeader, uncleHeader.mixHash(), Nonce(uncleHeaderObj["nonce"].get_str()));
+
+		if (overwrite == "mixHash")
+			updateEthashSeal(uncleHeader, h256(uncleHeaderObj["mixHash"].get_str()), uncleHeader.nonce());
+
+		uncle.setBlockHeader(uncleHeader, RecalcBlockHeader::Verify);
+	}
+	else
+		uncle.setBlockHeader(uncleHeader, RecalcBlockHeader::UpdateAndVerify);
 }
 
-void updatePoW(BlockHeader& _bi)
+void compareBlocks(TestBlock const& _a, TestBlock const& _b)
 {
-	mine(_bi);
-	_bi.noteDirty();
+	if (sha3(RLP(_a.getBytes())[0].data()) != sha3(RLP(_b.getBytes())[0].data()))
+	{
+		cnote << "block header mismatch\n";
+		cnote << toHex(RLP(_a.getBytes())[0].data()) << "vs" << toHex(RLP(_b.getBytes())[0].data());
+	}
+
+	if (sha3(RLP(_a.getBytes())[1].data()) != sha3(RLP(_b.getBytes())[1].data()))
+		cnote << "txs mismatch\n";
+
+	if (sha3(RLP(_a.getBytes())[2].data()) != sha3(RLP(_b.getBytes())[2].data()))
+		cnote << "uncle list mismatch\n" << RLP(_a.getBytes())[2].data() << "\n" << RLP(_b.getBytes())[2].data();
 }
 
-mArray writeTransactionsToJson(Transactions const& _txs)
+mArray writeTransactionsToJson(TransactionQueue const& _txsQueue)
 {
+	Transactions const& txs = _txsQueue.topTransactions(std::numeric_limits<unsigned>::max());
 	mArray txArray;
-	for (auto const& txi: _txs)
+	for (auto const& txi: txs)
 	{
 		mObject txObject = fillJsonWithTransaction(txi);
 		txArray.push_back(txObject);
@@ -756,53 +507,96 @@ mArray writeTransactionsToJson(Transactions const& _txs)
 	return txArray;
 }
 
-mObject writeBlockHeaderToJson(mObject& _o, BlockHeader const& _bi)
+mObject writeBlockHeaderToJson(Ethash::BlockHeader const& _bi)
 {
-	_o["parentHash"] = toString(_bi.parentHash());
-	_o["uncleHash"] = toString(_bi.sha3Uncles());
-	_o["coinbase"] = toString(_bi.beneficiary());
-	_o["stateRoot"] = toString(_bi.stateRoot());
-	_o["transactionsTrie"] = toString(_bi.transactionsRoot());
-	_o["receiptTrie"] = toString(_bi.receiptsRoot());
-	_o["bloom"] = toString(_bi.logBloom());
-	_o["difficulty"] = toCompactHex(_bi.difficulty(), HexPrefix::Add, 1);
-	_o["number"] = toCompactHex(_bi.number(), HexPrefix::Add, 1);
-	_o["gasLimit"] = toCompactHex(_bi.gasLimit(), HexPrefix::Add, 1);
-	_o["gasUsed"] = toCompactHex(_bi.gasUsed(), HexPrefix::Add, 1);
-	_o["timestamp"] = toCompactHex(_bi.timestamp(), HexPrefix::Add, 1);
-	_o["extraData"] = toHex(_bi.extraData(), 2, HexPrefix::Add);
-	_o["mixHash"] = toString(_bi.mixHash());
-	_o["nonce"] = toString(_bi.nonce());
-	_o["hash"] = toString(_bi.hash());
-	return _o;
+	mObject o;
+	o["parentHash"] = toString(_bi.parentHash());
+	o["uncleHash"] = toString(_bi.sha3Uncles());
+	o["coinbase"] = toString(_bi.beneficiary());
+	o["stateRoot"] = toString(_bi.stateRoot());
+	o["transactionsTrie"] = toString(_bi.transactionsRoot());
+	o["receiptTrie"] = toString(_bi.receiptsRoot());
+	o["bloom"] = toString(_bi.logBloom());
+	o["difficulty"] = toCompactHex(_bi.difficulty(), HexPrefix::Add, 1);
+	o["number"] = toCompactHex(_bi.number(), HexPrefix::Add, 1);
+	o["gasLimit"] = toCompactHex(_bi.gasLimit(), HexPrefix::Add, 1);
+	o["gasUsed"] = toCompactHex(_bi.gasUsed(), HexPrefix::Add, 1);
+	o["timestamp"] = toCompactHex(_bi.timestamp(), HexPrefix::Add, 1);
+	o["extraData"] = toHex(_bi.extraData(), 2, HexPrefix::Add);
+	o["mixHash"] = toString(_bi.mixHash());
+	o["nonce"] = toString(_bi.nonce());
+	o["hash"] = toString(_bi.hash());
+	return o;
 }
 
-RLPStream createFullBlockFromHeader(BlockHeader const& _bi, bytes const& _txs, bytes const& _uncles)
+void checkJsonSectionForInvalidBlock(mObject& _blObj)
 {
-	RLPStream rlpStream;
-	_bi.streamRLP(rlpStream, WithProof);
-
-	RLPStream ret(3);
-	ret.appendRaw(rlpStream.out());
-	ret.appendRaw(_txs);
-	ret.appendRaw(_uncles);
-
-	return ret;
+	BOOST_CHECK(_blObj.count("blockHeader") == 0);
+	BOOST_CHECK(_blObj.count("transactions") == 0);
+	BOOST_CHECK(_blObj.count("uncleHeaders") == 0);
 }
 
-} }// Namespace Close
+void checkBlocks(TestBlock const& _blockFromFields, TestBlock const& _blockFromRlp, string const& _testname)
+{
+	Ethash::BlockHeader const& blockHeaderFromFields = _blockFromFields.getBlockHeader();
+	Ethash::BlockHeader const& blockFromRlp = _blockFromRlp.getBlockHeader();
+
+	BOOST_CHECK_MESSAGE(blockHeaderFromFields.headerHash(WithProof) == blockFromRlp.headerHash(WithProof), _testname + "hash in given RLP not matching the block hash!");
+	BOOST_CHECK_MESSAGE(blockHeaderFromFields.parentHash() == blockFromRlp.parentHash(), _testname + "parentHash in given RLP not matching the block parentHash!");
+	BOOST_CHECK_MESSAGE(blockHeaderFromFields.sha3Uncles() == blockFromRlp.sha3Uncles(), _testname + "sha3Uncles in given RLP not matching the block sha3Uncles!");
+	BOOST_CHECK_MESSAGE(blockHeaderFromFields.beneficiary() == blockFromRlp.beneficiary(), _testname + "beneficiary in given RLP not matching the block beneficiary!");
+	BOOST_CHECK_MESSAGE(blockHeaderFromFields.stateRoot() == blockFromRlp.stateRoot(), _testname + "stateRoot in given RLP not matching the block stateRoot!");
+	BOOST_CHECK_MESSAGE(blockHeaderFromFields.transactionsRoot() == blockFromRlp.transactionsRoot(), _testname + "transactionsRoot in given RLP not matching the block transactionsRoot!");
+	BOOST_CHECK_MESSAGE(blockHeaderFromFields.receiptsRoot() == blockFromRlp.receiptsRoot(), _testname + "receiptsRoot in given RLP not matching the block receiptsRoot!");
+	BOOST_CHECK_MESSAGE(blockHeaderFromFields.logBloom() == blockFromRlp.logBloom(), _testname + "logBloom in given RLP not matching the block logBloom!");
+	BOOST_CHECK_MESSAGE(blockHeaderFromFields.difficulty() == blockFromRlp.difficulty(), _testname + "difficulty in given RLP not matching the block difficulty!");
+	BOOST_CHECK_MESSAGE(blockHeaderFromFields.number() == blockFromRlp.number(), _testname + "number in given RLP not matching the block number!");
+	BOOST_CHECK_MESSAGE(blockHeaderFromFields.gasLimit() == blockFromRlp.gasLimit(),"testname + gasLimit in given RLP not matching the block gasLimit!");
+	BOOST_CHECK_MESSAGE(blockHeaderFromFields.gasUsed() == blockFromRlp.gasUsed(), _testname + "gasUsed in given RLP not matching the block gasUsed!");
+	BOOST_CHECK_MESSAGE(blockHeaderFromFields.timestamp() == blockFromRlp.timestamp(), _testname + "timestamp in given RLP not matching the block timestamp!");
+	BOOST_CHECK_MESSAGE(blockHeaderFromFields.extraData() == blockFromRlp.extraData(), _testname + "extraData in given RLP not matching the block extraData!");
+	BOOST_CHECK_MESSAGE(blockHeaderFromFields.mixHash() == blockFromRlp.mixHash(), _testname + "mixHash in given RLP not matching the block mixHash!");
+	BOOST_CHECK_MESSAGE(blockHeaderFromFields.nonce() == blockFromRlp.nonce(), _testname + "nonce in given RLP not matching the block nonce!");
+
+	BOOST_CHECK_MESSAGE(blockHeaderFromFields == blockFromRlp, _testname + "However, blockHeaderFromFields != blockFromRlp!");
+
+	vector<TestTransaction> const& txsFromField = _blockFromFields.getTestTransactions();
+	vector<TestTransaction> const& txsFromRlp = _blockFromRlp.getTestTransactions();
+	BOOST_CHECK_MESSAGE(txsFromRlp.size() == txsFromField.size(), _testname + "transaction list size does not match");
+
+	for (size_t i = 0; i < txsFromField.size(); ++i)
+	{
+		Transaction const& trField = txsFromField.at(i).getTransaction();
+		Transaction const& trRlp = txsFromRlp.at(i).getTransaction();
+
+		BOOST_CHECK_MESSAGE(trField.data() == trRlp.data(), _testname + "transaction data in rlp and in field do not match");
+		BOOST_CHECK_MESSAGE(trField.gas() == trRlp.gas(), _testname + "transaction gasLimit in rlp and in field do not match");
+		BOOST_CHECK_MESSAGE(trField.gasPrice() == trRlp.gasPrice(), _testname + "transaction gasPrice in rlp and in field do not match");
+		BOOST_CHECK_MESSAGE(trField.nonce() == trRlp.nonce(), _testname + "transaction nonce in rlp and in field do not match");
+		BOOST_CHECK_MESSAGE(trField.signature().r == trRlp.signature().r, _testname + "transaction r in rlp and in field do not match");
+		BOOST_CHECK_MESSAGE(trField.signature().s == trRlp.signature().s, _testname + "transaction s in rlp and in field do not match");
+		BOOST_CHECK_MESSAGE(trField.signature().v == trRlp.signature().v, _testname + "transaction v in rlp and in field do not match");
+		BOOST_CHECK_MESSAGE(trField.receiveAddress() == trRlp.receiveAddress(), _testname + "transaction receiveAddress in rlp and in field do not match");
+		BOOST_CHECK_MESSAGE(trField.value() == trRlp.value(), _testname + "transaction receiveAddress in rlp and in field do not match");
+
+		BOOST_CHECK_MESSAGE(trField == trRlp, _testname + "transactions from  rlp and transaction from field do not match");
+		BOOST_CHECK_MESSAGE(trField.rlp() == trRlp.rlp(), _testname + "transactions rlp do not match");
+	}
+
+	vector<TestBlock> const& unclesFromField = _blockFromFields.getUncles();
+	vector<TestBlock> const& unclesFromRlp = _blockFromRlp.getUncles();
+	BOOST_REQUIRE_EQUAL(unclesFromField.size(), unclesFromRlp.size());
+	for (size_t i = 0; i < unclesFromField.size(); ++i)
+		BOOST_CHECK_MESSAGE(unclesFromField.at(i).getBlockHeader() == unclesFromRlp.at(i).getBlockHeader(), _testname + "block header in rlp and in field do not match at uncles");
+}
+
+}}//namespaces
 
 BOOST_AUTO_TEST_SUITE(BlockChainTests)
 
-BOOST_AUTO_TEST_CASE(bcRandom)
-{
-	dev::test::executeTests("bl201507071825GO", "/BlockchainTests/RandomTests",dev::test::getFolder(__FILE__) + "/BlockchainTestsFiller", dev::test::doBlockchainTests);
-}
-
 BOOST_AUTO_TEST_CASE(bcForkBlockTest)
 {
-	if (!test::Options::get().fillTests)
-		dev::test::executeTests("bcForkBlockTest", "/BlockchainTests",dev::test::getFolder(__FILE__) + "/BlockchainTestsFiller", dev::test::doBlockchainTests);
+	dev::test::executeTests("bcForkBlockTest", "/BlockchainTests",dev::test::getFolder(__FILE__) + "/BlockchainTestsFiller", dev::test::doBlockchainTests);
 }
 
 BOOST_AUTO_TEST_CASE(bcTotalDifficultyTest)
@@ -810,16 +604,9 @@ BOOST_AUTO_TEST_CASE(bcTotalDifficultyTest)
 	dev::test::executeTests("bcTotalDifficultyTest", "/BlockchainTests",dev::test::getFolder(__FILE__) + "/BlockchainTestsFiller", dev::test::doBlockchainTests);
 }
 
-BOOST_AUTO_TEST_CASE(bcForkUncleTest)
-{
-	if (!test::Options::get().fillTests)
-		dev::test::executeTests("bcForkUncle", "/BlockchainTests",dev::test::getFolder(__FILE__) + "/BlockchainTestsFiller", dev::test::doBlockchainTests);
-}
-
 BOOST_AUTO_TEST_CASE(bcInvalidRLPTest)
 {
-	if (!test::Options::get().fillTests)
-		dev::test::executeTests("bcInvalidRLPTest", "/BlockchainTests",dev::test::getFolder(__FILE__) + "/BlockchainTestsFiller", dev::test::doBlockchainTests);
+	dev::test::executeTests("bcInvalidRLPTest", "/BlockchainTests",dev::test::getFolder(__FILE__) + "/BlockchainTestsFiller", dev::test::doBlockchainTests);
 }
 
 BOOST_AUTO_TEST_CASE(bcRPC_API_Test)
